Scaffold tool pages, add same functionalities as yarn

+++ Models/Tool.cs
@@ -5,7 +5,7 @@
         public string Material { get; set; }
         public string Type { get; set; }
         public string Brand { get; set; }
-        public int Size { get; set; }
+        public float Size { get; set; }
         public ICollection<Pattern> Patterns { get; set; }
     }
 }
